+++ apps/web/app/future/apps/installed/[category]/page.tsx
@@ -1,4 +1,3 @@
-import Page from "@pages/apps/installed/[category]";
 import { withAppDirSsr } from "app/WithAppDirSsr";
 import { _generateMetadata } from "app/_utils";
 import { WithLayout } from "app/layoutHOC";
@@ -6,6 +5,8 @@ import { WithLayout } from "app/layoutHOC";
 import { APP_NAME } from "@calcom/lib/constants";
 
 import { getServerSideProps } from "@lib/apps/installed/[category]/getServerSideProps";
+
+import Page from "@components/pages/apps/installed/[category]";
 
 export const generateMetadata = async () => {
   return await _generateMetadata(
